+++ pyro/poutine/broadcast_messenger.py
@@ -5,11 +5,32 @@ from .messenger import Messenger
 
 class BroadcastMessenger(Messenger):
     """
-    `BroadcastMessenger` automatically broadcasts the batch shape of
-    the stochastic function at a sample site when inside a single
-    or nested iarange context. The existing `batch_shape` must be
-    broadcastable with the size of the :class:`~pyro.iarange`
-    contexts installed in the `cond_indep_stack`.
+    Automatically broadcasts the batch shape of the stochastic function
+    at a sample site when inside a single or nested iarange context.
+    The existing `batch_shape` must be broadcastable with the size
+    of the :class:`~pyro.iarange` contexts installed in the
+    `cond_indep_stack`.
+
+    Notice how `model_automatic_broadcast` below automates expanding of
+    distribution batch shapes. This makes it easy to modularize a
+    Pyro model as the sub-components are agnostic of the wrapping
+    :class:`~pyro.iarange` contexts.
+
+    >>> def model_broadcast_by_hand():
+    ...     with pyro.iarange("batch", 100, dim=-2):
+    ...         with pyro.iarange("components", 3, dim=-1):
+    ...             sample = pyro.sample("sample", dist.Bernoulli(torch.ones(3) * 0.5)
+    ...                                                .expand_by(100))
+    ...             assert sample.shape == torch.Size((100, 3))
+    ...     return sample
+
+    >>> @poutine.broadcast
+    ... def model_automatic_broadcast():
+    ...     with pyro.iarange("batch", 100, dim=-2):
+    ...         with pyro.iarange("components", 3, dim=-1):
+    ...             sample = pyro.sample("sample", dist.Bernoulli(torch.tensor(0.5)))
+    ...             assert sample.shape == torch.Size((100, 3))
+    ...     return sample
     """
     def _pyro_sample(self, msg):
         """
