+++ pyro/poutine/trace_messenger.py
@@ -31,14 +31,28 @@ def identify_dense_edges(trace):
 
 class TraceMessenger(Messenger):
     """
-    Execution trace messenger.
+    Return a handler that records the inputs and outputs of primitive calls
+    and their dependencies.
 
-    A TraceMessenger records the input and output to every Pyro primitive
-    and stores them as a site in a Trace().
-    This should, in theory, be sufficient information for every inference algorithm
-    (along with the implicit computational graph in the Variables?)
+    Consider the following Pyro program:
 
-    We can also use this for visualization.
+        >>> def model(x):
+        ...     s = pyro.param("s", torch.tensor(0.5))
+        ...     z = pyro.sample("z", dist.Normal(x, s))
+        ...     return z ** 2
+
+    We can record its execution using ``trace``
+    and use the resulting data structure to compute the log-joint probability
+    of all of the sample sites in the execution or extract all parameters.
+
+        >>> trace = trace(model).get_trace(0.0)
+        >>> logp = trace.log_prob_sum()
+        >>> params = [trace.nodes[name]["value"].unconstrained() for name in trace.param_nodes]
+
+    :param fn: a stochastic function (callable containing Pyro primitive calls)
+    :param graph_type: string that specifies the kind of graph to construct
+    :param param_only: if true, only records params and not samples
+    :returns: stochastic function decorated with a :class:`~pyro.poutine.trace_messenger.TraceMessenger`
     """
 
     def __init__(self, graph_type=None, param_only=None, strict_names=None):
